Add files via upload
Braces in a string are considered to be balanced if the following criteria are met:
All braces must be closed. Braces come in pairs of the form (), {} and [].
The left brace opens the pair, and the right one closes it.
In any set of nested braces, the braces between any pair must be closed.
For example, [{}] is a valid grouping of braces but [}]{} is not.